CCHS-384: Updated data domain table truncation to only occur when the merged item has a data domain in the table.

+++ coastal-hazards-portal/src/main/java/gov/usgs/cida/coastalhazards/jpa/ItemManager.java
@@ -164,10 +164,15 @@ public class ItemManager implements AutoCloseable {
 			}
 			item.setChildren(replaceList);
 		}
-		em.merge(item);
+		em.merge(item);	
+		id = item.getId();	
 		DataDomainManager dm = new DataDomainManager();
-		dm.deleteAll();
-		id = item.getId();
+		
+		//Clear the data domains if this item has a domain
+		if(dm.load(id) != null){
+		    dm.deleteAll();
+		}
+		
 		return id;
 	}
